minor fixes

+++ app.py
@@ -141,12 +141,12 @@ def database_findProject(project_name):
     return projects if projects else []
 def database_deleteItem(item_id):
     supply = db.session.get(Supply, item_id)
-    if supply:
+    if supply and supply.owner == current_user.id:
         db.session.delete(supply)
         db.session.commit()
 def database_deleteProject(project_id):
     project = db.session.get(Project, project_id)
-    if project:
+    if project and project.owner == current_user.id:
         db.session.delete(project)
         db.session.commit()
 def database_deleteSupplyImage(image_path):
@@ -161,7 +161,7 @@ def database_deleteProjectImage(image_path):
         db.session.commit()
 def database_deleteProjectItem(project_item_id):
     project_item = db.session.get(ProjectItem, project_item_id)
-    if project_item:
+    if project_item and project_item.project.owner == current_user.id:
         db.session.delete(project_item)
         db.session.commit()
 def database_multiparameter_item_search(search_terms):
